Renamed App component to Weather
